move common functions into common dir
